Move some things around, consolidate to remove circular dependencies
