+++ src/schemaFromIntrospectionData.js
@@ -1,4 +1,8 @@
 // @flow
+/**
+ * Takes the introspectionQuery response and parses it into the "Schema"
+ * type that we use to look up types, interfaces, etc.
+ */
 import type {IntrospectionQuery} from 'graphql';
 import type {Schema} from './types';
